requirements.txt
Flask==3.0.1
google_api_python_client==2.115.0
google_auth_oauthlib==1.2.0
isodate==0.6.1
pandas==2.2.0
protobuf==4.25.2
Requests==2.31.0
SQLAlchemy==1.3.19
Pyarrow==15.0.0
langdetect==1.0.9
gunicorn==20.1.0